refactor(editor): extract shared Composition module, wire into SemanticWindow builder
Extracts the text composition pipeline (conceal application, inline
virtual text injection) from Renderer.Line into a shared
Renderer.Composition module. Both the draw path (Line.ex) and the
semantic path (SemanticWindow.Builder) now call the same composition
functions, ensuring identical output for the same input.

This replaces the compose_line_text stub in Builder.ex with the real
composition pipeline. The semantic path now correctly handles:
- Inline virtual text (LSP inlay hints, ghost text) spliced at anchor
  columns
- Conceal ranges (hidden import paths, collapsed markdown links) with
  replacement characters
- Decoration highlight merging (search matches, etc.)

New module:
- Renderer.Composition: apply_conceals/3, inject_inline_virtual_text/3,
  compose_segments/3, segments_to_text_and_spans/1,
  split_text_at_display_col/2

Changes:
- Line.ex delegates to Composition public functions (pure refactor,
  no logic changes in the moved functions)
- Builder.ex replaces compose_line_text stub + build_spans_for_line
  with compose_line/4 that runs the full pipeline:
  highlight segments -> merge decorations -> apply conceals ->
  inject virtual text -> convert to text + spans

The pair-based conceal functions (used by selection rendering) remain
in Line.ex as they're not needed by the semantic path.
