+++ lib/minga/editor/renderer/line.ex
@@ -18,6 +18,7 @@ defmodule Minga.Editor.Renderer.Line do
   alias Minga.Buffer.Decorations.ConcealRange
   alias Minga.Buffer.Unicode
   alias Minga.Editor.DisplayList
+  alias Minga.Editor.Renderer.Composition
   alias Minga.Editor.Renderer.Context
   alias Minga.Face
   alias Minga.Highlight
@@ -339,137 +340,9 @@ defmodule Minga.Editor.Renderer.Line do
   defp render_decorated_plain_line(line_text, screen_row, buf_line, ctx, line_highlights) do
     segments = [{line_text, Face.new()}]
     segments = Decorations.merge_highlights(segments, line_highlights, buf_line)
-    segments = apply_conceals_to_segments(segments, ctx.decorations, buf_line)
-    segments = inject_inline_virtual_text(segments, ctx.decorations, buf_line)
+    segments = Composition.apply_conceals(segments, ctx.decorations, buf_line)
+    segments = Composition.inject_inline_virtual_text(segments, ctx.decorations, buf_line)
     render_segments_with_scroll(segments, screen_row, ctx)
-  end
-
-  # ── Inline virtual text injection ────────────────────────────────────────────
-
-  # Injects inline virtual text segments into the styled segment list at their
-  # anchor column positions, displacing subsequent content rightward.
-  @spec inject_inline_virtual_text(
-          [{String.t(), Face.t()}],
-          Decorations.t(),
-          non_neg_integer()
-        ) :: [{String.t(), Face.t()}]
-  defp inject_inline_virtual_text(segments, decorations, buf_line) do
-    inline_vts = Decorations.inline_virtual_texts_for_line(decorations, buf_line)
-
-    if inline_vts == [] do
-      segments
-    else
-      do_inject_inline(segments, inline_vts, 0, [])
-    end
-  end
-
-  # Walk through segments, tracking the current buffer column. When a virtual
-  # text's anchor column falls within the current segment, split the segment
-  # and insert the virtual text segments between the halves.
-  @spec do_inject_inline(
-          [{String.t(), Face.t()}],
-          [Decorations.VirtualText.t()],
-          non_neg_integer(),
-          [{String.t(), Face.t()}]
-        ) :: [{String.t(), Face.t()}]
-  defp do_inject_inline([], remaining_vts, _col, acc) do
-    # Append any remaining virtual texts after all buffer content
-    vt_segments = Enum.flat_map(remaining_vts, fn vt -> vt.segments end)
-    Enum.reverse(acc, vt_segments)
-  end
-
-  defp do_inject_inline(segments, [], _col, acc) do
-    # No more virtual texts to inject, append remaining segments
-    Enum.reverse(acc, segments)
-  end
-
-  defp do_inject_inline(
-         [{seg_text, seg_style} | rest_segs],
-         [%{anchor: {_l, anchor_col}} = vt | rest_vts],
-         col,
-         acc
-       ) do
-    seg_width = Unicode.display_width(seg_text)
-    seg_end = col + seg_width
-
-    inject_at_position(
-      {seg_text, seg_style},
-      rest_segs,
-      vt,
-      rest_vts,
-      col,
-      seg_end,
-      anchor_col,
-      acc
-    )
-  end
-
-  # Virtual text anchor is at or before the current segment start: inject before
-  @spec inject_at_position(
-          {String.t(), Face.t()},
-          [{String.t(), Face.t()}],
-          Decorations.VirtualText.t(),
-          [Decorations.VirtualText.t()],
-          non_neg_integer(),
-          non_neg_integer(),
-          non_neg_integer(),
-          [{String.t(), Face.t()}]
-        ) :: [{String.t(), Face.t()}]
-  defp inject_at_position(seg, rest_segs, vt, rest_vts, col, _seg_end, anchor_col, acc)
-       when anchor_col <= col do
-    do_inject_inline([seg | rest_segs], rest_vts, col, vt.segments ++ acc)
-  end
-
-  # Virtual text anchor is within the current segment: split and inject
-  defp inject_at_position(
-         {seg_text, seg_style},
-         rest_segs,
-         vt,
-         rest_vts,
-         col,
-         seg_end,
-         anchor_col,
-         acc
-       )
-       when anchor_col < seg_end do
-    split_at = anchor_col - col
-    {before_text, after_text} = split_text_at_display_col(seg_text, split_at)
-    after_part = if after_text != "", do: [{after_text, seg_style}], else: []
-    new_acc = after_part ++ vt.segments ++ [{before_text, seg_style} | acc]
-    do_inject_inline(rest_segs, rest_vts, seg_end, new_acc)
-  end
-
-  # Virtual text anchor is after the current segment: skip segment, keep going
-  defp inject_at_position(
-         {seg_text, seg_style},
-         rest_segs,
-         vt,
-         rest_vts,
-         _col,
-         seg_end,
-         _anchor_col,
-         acc
-       ) do
-    do_inject_inline(rest_segs, [vt | rest_vts], seg_end, [{seg_text, seg_style} | acc])
-  end
-
-  # Split text at a display column position (not byte or grapheme index).
-  @spec split_text_at_display_col(String.t(), non_neg_integer()) :: {String.t(), String.t()}
-  defp split_text_at_display_col(text, display_col) do
-    graphemes = String.graphemes(text)
-
-    {before_acc, after_acc, _} =
-      Enum.reduce(graphemes, {[], [], 0}, fn g, {bef, aft, col} ->
-        w = Unicode.grapheme_width(g)
-
-        if col < display_col do
-          {[g | bef], aft, col + w}
-        else
-          {bef, [g | aft], col + w}
-        end
-      end)
-
-    {before_acc |> Enum.reverse() |> Enum.join(), after_acc |> Enum.reverse() |> Enum.join()}
   end
 
   # ── Conceal application to grapheme pairs (for selection paths) ──────────────
@@ -564,235 +437,6 @@ defmodule Minga.Editor.Renderer.Line do
     [pair | do_apply_conceals_to_pairs(rest, conceals, line, col + w)]
   end
 
-  # ── Conceal range application ─────────────────────────────────────────────────
-
-  # Applies conceal ranges to a styled segment list. Concealed graphemes are
-  # removed from the output; if a conceal has a replacement character, that
-  # replacement is emitted once in place of the entire concealed range.
-  #
-  # Operates in buffer column space (before inline virtual text shifts).
-  # Walks the segment list tracking the current buffer column, and for each
-  # conceal range on this line, excises the concealed graphemes and optionally
-  # inserts the replacement.
-  @spec apply_conceals_to_segments(
-          [{String.t(), Face.t()}],
-          Decorations.t(),
-          non_neg_integer()
-        ) :: [{String.t(), Face.t()}]
-  defp apply_conceals_to_segments(segments, decorations, buf_line) do
-    conceals = Decorations.conceals_for_line(decorations, buf_line)
-
-    if conceals == [] do
-      segments
-    else
-      do_apply_conceals(segments, conceals, %{line: buf_line, col: 0, acc: []})
-    end
-  end
-
-  # Walk segments and conceals together. Both are sorted by column position.
-  # Uses a conceal_ctx map to bundle line/col/acc and avoid high arity helpers.
-  @typep conceal_ctx :: %{
-           line: non_neg_integer(),
-           col: non_neg_integer(),
-           acc: [{String.t(), Face.t()}]
-         }
-
-  # Dialyzer has trouble inferring Face.t() through the conceal pipeline's
-  # recursive segment splitting. The specs are correct but dialyzer's
-  # inference chain disagrees. Suppress until a deeper refactor resolves
-  # the type flow (tracked by the Face migration).
-  @dialyzer {:nowarn_function, do_apply_conceals: 3}
-  @dialyzer {:nowarn_function, apply_conceal_from_start: 7}
-  @dialyzer {:nowarn_function, apply_conceal_mid_segment: 7}
-  @dialyzer {:nowarn_function, maybe_emit_replacement: 5}
-
-  @spec do_apply_conceals(
-          [{String.t(), Face.t()}],
-          [ConcealRange.t()],
-          conceal_ctx()
-        ) :: [{String.t(), Face.t()}]
-  defp do_apply_conceals([], _conceals, ctx), do: Enum.reverse(ctx.acc)
-  defp do_apply_conceals(segments, [], ctx), do: Enum.reverse(ctx.acc, segments)
-
-  defp do_apply_conceals(
-         [{seg_text, seg_style} | rest_segs],
-         [%ConcealRange{} = conceal | rest_conceals] = conceals,
-         ctx
-       ) do
-    seg_width = Unicode.display_width(seg_text)
-    seg_end = ctx.col + seg_width
-
-    # Determine conceal start/end on this line
-    {_sl, sc} = conceal.start_pos
-    {el, ec} = conceal.end_pos
-    conceal_start = if elem(conceal.start_pos, 0) < ctx.line, do: 0, else: sc
-    conceal_end = if el > ctx.line, do: seg_end + 1, else: ec
-
-    # Dispatch based on the relationship between segment and conceal.
-    # Uses if/else instead of separate function to avoid high arity.
-    if conceal_start >= seg_end do
-      # Conceal is entirely past this segment: emit segment, keep going
-      do_apply_conceals(
-        rest_segs,
-        conceals,
-        %{ctx | col: seg_end, acc: [{seg_text, seg_style} | ctx.acc]}
-      )
-    else
-      if conceal_start <= ctx.col do
-        apply_conceal_from_start(
-          {seg_text, seg_style},
-          rest_segs,
-          conceal,
-          rest_conceals,
-          seg_end,
-          conceal_end,
-          ctx
-        )
-      else
-        apply_conceal_mid_segment(
-          {seg_text, seg_style},
-          rest_segs,
-          conceal,
-          rest_conceals,
-          seg_end,
-          conceal_start,
-          ctx
-        )
-      end
-    end
-  end
-
-  # Conceal starts at or before segment start: skip concealed portion
-  @spec apply_conceal_from_start(
-          {String.t(), Face.t()},
-          [{String.t(), Face.t()}],
-          ConcealRange.t(),
-          [ConcealRange.t()],
-          non_neg_integer(),
-          non_neg_integer(),
-          conceal_ctx()
-        ) :: [{String.t(), Face.t()}]
-  defp apply_conceal_from_start(
-         {seg_text, seg_style},
-         rest_segs,
-         conceal,
-         rest_conceals,
-         seg_end,
-         conceal_end,
-         ctx
-       ) do
-    if conceal_end >= seg_end do
-      # Entire segment is concealed. Emit replacement on first concealed segment
-      # only (when col == conceal_start or conceal started on a previous segment).
-      replacement_acc = maybe_emit_replacement(ctx.acc, conceal, ctx.col, ctx.line, seg_style)
-
-      if conceal_end == seg_end do
-        # Conceal ends exactly at segment end: move to next conceal
-        do_apply_conceals(rest_segs, rest_conceals, %{ctx | col: seg_end, acc: replacement_acc})
-      else
-        # Conceal extends past this segment: continue with same conceal
-        do_apply_conceals(
-          rest_segs,
-          [conceal | rest_conceals],
-          %{ctx | col: seg_end, acc: replacement_acc}
-        )
-      end
-    else
-      # Conceal ends within this segment: split and emit the non-concealed tail
-      replacement_acc = maybe_emit_replacement(ctx.acc, conceal, ctx.col, ctx.line, seg_style)
-      drop_cols = conceal_end - ctx.col
-      {_before, after_text} = split_text_at_display_col(seg_text, drop_cols)
-
-      if after_text != "" do
-        do_apply_conceals(
-          [{after_text, seg_style} | rest_segs],
-          rest_conceals,
-          %{ctx | col: conceal_end, acc: replacement_acc}
-        )
-      else
-        do_apply_conceals(
-          rest_segs,
-          rest_conceals,
-          %{ctx | col: seg_end, acc: replacement_acc}
-        )
-      end
-    end
-  end
-
-  # Conceal starts within this segment: emit the before portion, then handle concealed part
-  @spec apply_conceal_mid_segment(
-          {String.t(), Face.t()},
-          [{String.t(), Face.t()}],
-          ConcealRange.t(),
-          [ConcealRange.t()],
-          non_neg_integer(),
-          non_neg_integer(),
-          conceal_ctx()
-        ) :: [{String.t(), Face.t()}]
-  defp apply_conceal_mid_segment(
-         {seg_text, seg_style},
-         rest_segs,
-         conceal,
-         rest_conceals,
-         seg_end,
-         conceal_start,
-         ctx
-       ) do
-    split_at = conceal_start - ctx.col
-    {before_text, after_text} = split_text_at_display_col(seg_text, split_at)
-
-    before_acc = if before_text != "", do: [{before_text, seg_style} | ctx.acc], else: ctx.acc
-
-    if after_text != "" do
-      # Re-process the remainder (which now starts at the conceal)
-      do_apply_conceals(
-        [{after_text, seg_style} | rest_segs],
-        [conceal | rest_conceals],
-        %{ctx | col: conceal_start, acc: before_acc}
-      )
-    else
-      # Before text consumed everything (edge case with wide chars)
-      do_apply_conceals(
-        rest_segs,
-        [conceal | rest_conceals],
-        %{ctx | col: seg_end, acc: before_acc}
-      )
-    end
-  end
-
-  # Emits the replacement character for a conceal range, but only once
-  # (at the start of the conceal). We detect "first concealed segment"
-  # by checking if we're at the conceal's start column on this line.
-  #
-  # The seg_style from the concealed text is merged onto the replacement
-  # style so that overlapping decorations (e.g., search highlight bg)
-  # carry through to the replacement character.
-  @spec maybe_emit_replacement(
-          [{String.t(), Face.t()}],
-          ConcealRange.t(),
-          non_neg_integer(),
-          non_neg_integer(),
-          Face.t()
-        ) :: [{String.t(), Face.t()}]
-  defp maybe_emit_replacement(acc, %ConcealRange{replacement: nil}, _col, _line, _seg_style),
-    do: acc
-
-  defp maybe_emit_replacement(acc, conceal, col, line, seg_style) do
-    {sl, sc} = conceal.start_pos
-    conceal_start_on_line = if sl < line, do: 0, else: sc
-
-    if col <= conceal_start_on_line do
-      # Merge: conceal replacement_style wins over the segment's style,
-      # but the segment's style provides bg/fg from overlapping decorations.
-      merged_style =
-        Decorations.merge_style_props(seg_style, conceal.replacement_style)
-
-      [{conceal.replacement, merged_style} | acc]
-    else
-      acc
-    end
-  end
-
   # ── Syntax highlighting ──────────────────────────────────────────────────────
 
   @spec render_highlighted_line(
@@ -820,8 +464,8 @@ defmodule Minga.Editor.Renderer.Line do
 
     # Merge decoration highlight ranges with syntax segments (pre-queried, no double lookup)
     segments = Decorations.merge_highlights(segments, line_highlights, buf_line)
-    segments = apply_conceals_to_segments(segments, ctx.decorations, buf_line)
-    segments = inject_inline_virtual_text(segments, ctx.decorations, buf_line)
+    segments = Composition.apply_conceals(segments, ctx.decorations, buf_line)
+    segments = Composition.inject_inline_virtual_text(segments, ctx.decorations, buf_line)
 
     render_segments_with_scroll(segments, screen_row, ctx)
   end
